requirements.txt
spidev
Pillow
zhdate
requests
psutil
urllib3
watchdog
numpy
pyyaml
